ram_optimize_final-main/requirements.txt
playwright
cryptography
rich
psutil
pystray
pillow
flask
flask-cors
